+++ .env.local
@@ -12,11 +12,12 @@ CONTAINER_REGISTRY=local
 
 # Backend variables.
 BACKEND_VERSION=v0.0.1
-RABBITMQ_BROKER_ADDRESS=rabbitmq
+RABBITMQ_BROKER_ADDRESS=localhost
 RABBITMQ_BROKER_PORT=5672
 
 # Frontend variables.
 FRONTEND_VERSION=v0.0.1
+VITE_BACKEND_URL=http://localhost:8888
 
 # Frontend proxy variables.
 FRONTEND_PROXY_VERSION=v0.0.1
