--- backend/entrypoint.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env sh
-
-set -e
-
-python -m hades.main server
