+++ backend/hades/messages/rabbitmq/exchange.py
@@ -1,7 +1,6 @@
 """Defines a RabbitMQ exchange."""
 
 # Standard library imports.
-from os import environ
 from typing import Callable
 
 # Third-party imports.
@@ -14,32 +13,39 @@ from .broker import RabbitMQBroker
 class RabbitMQExchange:
     def __init__(
       self,
-      name: str,
       broker: RabbitMQBroker,
-      handler: Callable
+      virtual_host: str,
+      username: str,
+      password: str,
+      name: str,
+      routing_key: str,
+      handler: Callable,
     ):
-        self.name = name
-        self.routing_key = self.name.split(".")[-1]
         self.parameters = ConnectionParameters(
             host=broker.address,
             port=broker.port,
-            virtual_host="/",
+            virtual_host=virtual_host,
             credentials=PlainCredentials(
-                username=environ["RABBITMQ_USERNAME"],
-                password=environ["RABBITMQ_PASSWORD"]
+                username=username,
+                password=password
             )
         )
         self.connection = BlockingConnection(self.parameters)
         self.channel = self.connection.channel()
+        self.name = name
         self.channel.exchange_declare(
+            exchange=self.name,
             exchange_type="topic",
-            exchange=self.name
         )
-        self.result = self.channel.queue_declare(queue=f"{self.name}.queue", durable=True)
+        self.result = self.channel.queue_declare(
+            queue=f"{self.name}.queue",
+            durable=False
+        )
         self.queue = self.result.method.queue
+        self.routing_key = routing_key
         self.channel.queue_bind(
             exchange=self.name,
             queue=self.queue,
-            routing_key=self.routing_key
+            routing_key=routing_key
         )
         self.handler = handler
