+++ backend/hades/server/__init__.py
@@ -1,1 +1,1 @@
-from .server import HadesServer
+from .server import HadesServer, Start
