+++ frontend/hades/src/pages/home.tsx
@@ -1,8 +1,7 @@
-import * as React from 'react';
 import Typography from '@mui/material/Typography';
 import { PageContainer } from '@toolpad/core/PageContainer';
 
-export default function HomePage() {
+export default function Home() {
   
 
   return (
