+++ rabbitmq/requirements.txt
@@ -1,0 +1,1 @@
+pika
